// my_client/src/app/models/album.ts
export class Album {
    _id: any;
    imgalbum?: string;
    imgalbum1?: string;
    imgalbum2?: string;
    imgalbum3?: string;
    imgalbum4?: string;
    imgalbum5?: string;
    imgalbum6?: string;
    imgalbum7?: string;
    imgalbum8?: string;
    imgalbum9?: string;
    imgalbum10?: string;
    name?: string;
    quantity?: number;
  }
  